rename: vote 파일 구조 변경 (승휘님 리뷰 반영)
