+++ src/App.tsx
@@ -3,7 +3,7 @@ import { PageEndpoints } from "@/constants/endpoints";
 import PromotionMain from "@/pages/promotions/PromotionMain";
 import PromotionDetail from "@/pages/promotions/detail/PromotionDetail";
 import Home from "@/pages/home/Home";
-import Vote from "@/pages/vote/Vote";
+import Vote from "@/pages/vote/select/Vote";
 import VoteResult from "@/pages/vote/result/VoteResult";
 import SignIn from "@/pages/auth/signIn/SignIn";
 import SignUp from "@/pages/auth/signUp/SignUp";
